Add login controller to the ClientAPI
- Basic login controller to get past the login screen
- Provide hard-coded character load out
