+++ ClientApi.Api/Models/CaisStatus.cs
@@ -1,0 +1,9 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class CaisStatus
+    {
+        public string State { get; set; }
+        public long Duration { get; set; }
+        public long ExpiresAt { get; set; }
+    }
+}
