+++ ClientApi.Api/Models/ColorItem.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class ColorItem : Item
+    {
+        public long Value { get; set; }
+    }
+}
