+++ ClientApi.Api/Models/ColorValue.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class ColorValue
+    {
+        public long Color { get; set; }
+    }
+}
