+++ ClientApi.Api/Models/ColoredItem.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class ColoredItem : Item
+    {
+        public ColorValue Value { get; set; }
+    }
+}
