+++ ClientApi.Api/Models/ColoredTransformableSdbItem.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class ColoredTransformableSdbItem : TransformableSdbItem
+    {
+        public long Color { get; set; }
+    }
+}
