+++ ClientApi.Api/Models/Gear.cs
@@ -1,0 +1,9 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class Gear
+    {
+        public int SlotTypeId { get; set; }
+        public int SdbId { get; set; }
+        public long ItemGuid { get; set; }
+    }
+}
