+++ ClientApi.Api/Models/HairItem.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class HairItem : Item
+    {
+        public ColorItem Color { get; set; }
+    }
+}
