+++ ClientApi.Api/Models/Item.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class Item
+    {
+        public int Id { get; set; }
+    }
+}
