+++ ClientApi.Api/Models/SdbItem.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class SdbItem
+    {
+        public int SdbId { get; set; }
+    }
+}
