+++ ClientApi.Api/Models/ServerList.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace ClientApi.Api.Models
+{
+    public class ServerList
+    {
+        public IEnumerable<long> ZoneList { get; set; }
+    }
+}
