+++ ClientApi.Api/Models/TransformableSdbItem.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace ClientApi.Api.Models
+{
+    public class TransformableSdbItem : SdbItem
+    {
+        public IEnumerable<decimal> Transform { get; set; }
+    }
+}
