+++ ClientApi.Api/Models/WarpaintPattern.cs
@@ -1,0 +1,7 @@
+﻿namespace ClientApi.Api.Models
+{
+    public class WarpaintPattern : TransformableSdbItem
+    {
+        public int Usage { get; set; }
+    }
+}
